ms-agent/requirements.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# HTTP requests для GigaChat API
requests==2.31.0

# Excel parsing
openpyxl==3.1.2
pandas==2.1.4

# Utilities
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Metrics
prometheus-client==0.19.0

# Type hints
typing-extensions==4.9.0

